+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Full documentation for RPP is available at [https://rocm.docs.amd.com/projects/r
 ## (Unreleased) RPP 2.0.0
 
 ### Added
-* RPP Bitwise NOT, Bitwise AND, Bitwise OR augmentations on HOST (CPU) and HIP backends. (#520)
+* Bitwise NOT, Bitwise AND, Bitwise OR augmentations on HOST (CPU) and HIP backends. (#520)
+* Tensor Concat augmentation on HOST (CPU) and HIP backends. (#530)
 * `log1p`, defined as `log (1 + x)`, tensor augmentation support on HOST (CPU) and HIP backends.
 
 ### Changed